training/requirements.txt
# Professional Drug Analysis Training Requirements
# Core ML/DL frameworks
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
tensorflowjs>=4.0.0

# Computer vision and image processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Object detection and YOLO
ultralytics>=8.0.0

# Hugging Face ecosystem
huggingface-hub>=0.16.0
transformers>=4.30.0
datasets>=2.13.0

# Experiment tracking and monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Data handling and utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0